[add] 集成mybatis-generator，生成user表的CRUD

+++ mnus-ucenter/src/main/java/com/mnus/ucenter/controller/UcenterController.java
@@ -18,7 +18,7 @@ public class UcenterController {
     private UcenterService ucenterService;
 
     @GetMapping("/count")
-    public Integer count() {
+    public Long count() {
         return ucenterService.count();
     }
 }
